requirements.txt

Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.0.1
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
psycopg2-binary==2.9.9
django-storages==1.14.2
boto3==1.34.0
twilio==8.10.0
sendgrid==6.10.0
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.3.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.25
qrcode==7.4.2
cryptography==41.0.8
